Endpoint refactoring to make less api calls and rating system overhaul
