+++ Server/CookEase.Api/Extensions/ServiceCollectionExtensions.cs
@@ -27,7 +27,6 @@ public static class ServiceCollectionExtensions
         services.AddScoped<IUserRepository, UserRepository>();
         services.AddScoped<IRecipeRepository, RecipeRepository>();
         services.AddScoped<IRecipeNutritionRepository, RecipeNutritionRepository>();
-        services.AddScoped<IRecipeRatingRepository, RecipeRatingRepository>();
         services.AddScoped<ICommentRepository, CommentRepository>();
         services.AddScoped<IRecipeCategoryRepository, RecipeCategoryRepository>();
