+++ Server/Infrastructure/AppDbContext.cs
@@ -19,8 +19,6 @@ public class AppDbContext(DbContextOptions<AppDbContext> options) : DbContext(op
 
     public required DbSet<RecipeNutrition> RecipeNutrition { get; set; }
 
-    public required DbSet<RecipeRating> RecipeRatings { get; set; }
-
     public required DbSet<Recipe> Recipes { get; set; }
 
     public required DbSet<User> Users { get; set; }
