+++ Server/Infrastructure/Repositories/CommentRepository.cs
@@ -27,11 +27,7 @@ public class CommentRepository : GenericRepository<Comment>, ICommentRepository
             .ToListAsync();
     }
 
-    public int GetCommentCountForRecipe(int recipeId)
-    {
-        return _comments.Count(x => x.RecipeId == recipeId);
-    }
-
+    // Implementation should be changed
     public async Task<Comment?> IncreaseCommentLikeCount(int commentId)
     {
         var comment = await _comments.FindAsync(commentId);
@@ -40,12 +36,13 @@ public class CommentRepository : GenericRepository<Comment>, ICommentRepository
             return null;
         }
 
-        comment.LikeCount++;
+        //comment.LikeCount++;
         _comments.Update(comment);
         await _context.SaveChangesAsync();
         return comment;
     }
 
+    // Implementation should be changed
     public async Task<Comment?> DecreaseCommentLikeCount(int commentId)
     {
         var comment = await _comments.FindAsync(commentId);
@@ -54,9 +51,41 @@ public class CommentRepository : GenericRepository<Comment>, ICommentRepository
             return null;
         }
 
-        comment.LikeCount--;
+        //comment.LikeCount--;
         _comments.Update(comment);
         await _context.SaveChangesAsync();
         return comment;
     }
+
+    public async Task<decimal> GetRecipeRating(int recipeId)
+    {
+        var recipeRatings = await _comments.Where(x => x.RecipeId == recipeId).ToListAsync();
+        if (recipeRatings.Count == 0)
+        {
+            return 0;
+        }
+
+        var averageRating = recipeRatings.Average(x => x.Rating);
+        return Math.Round(averageRating * 2, MidpointRounding.AwayFromZero) / 2;
+    }
+    
+    // Implementation should be changed
+    //public async Task UpdateUserRecipeRating(int userId, int recipeId, decimal newRatingValue)
+    //{
+    //    var recipeRating = await _recipeRatings.FindAsync(recipeId, userId);
+    //    if (recipeRating is null)
+    //    {
+    //        await Add(new RecipeRating
+    //        {
+    //            UserId = userId,
+    //            RecipeId = recipeId,
+    //            Rating = newRatingValue
+    //        });
+    //        return;
+    //    }
+
+    //    recipeRating.Rating = newRatingValue;
+    //    await Update(recipeRating);
+    //    await _context.SaveChangesAsync();
+    //}
 }
